pyproject.toml
[project]
name = "openvpn-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "domster704",email = "domster704@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standart] (>=0.116.2,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]

[tool.poetry]
packages = [
    { include = "src" },
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.1.0,<26.0.0)"
]
